deviding to folders
