+++ video_manipulation_functions/video_manipulation.py
@@ -1,10 +1,10 @@
-from flask import request, jsonify
 import cv2
 import shutil
 import os
-from BionicEye.given_functions import generate_metadata, is_frame_tagged
+from BionicEye.video_manipulation_functions.given_functions import generate_metadata, is_frame_tagged
 from BionicEye.app import Video, Metadata, Frame
-from BionicEye.manage_saving import OSManage, DBManage
+from BionicEye.singelton_classes.db_manager import DBManage
+from BionicEye.singelton_classes.os_manager import OSManage
 
 DB_MANAGE = DBManage()
 OS_MANAGE = OSManage()
@@ -95,21 +95,3 @@ def save_frames(video_path):
     OS_MANAGE.upload_dir(video_frames_dir)
     shutil.rmtree(all_frames_dir)
     shutil.rmtree(os.path.dirname(video_path))
-
-
-def add_video():
-    uploaded_file = request.files['file']
-    os.makedirs('videos', exist_ok=True)
-    videos_dir = os.path.join(os.getcwd(), 'videos')
-    video_path = os.path.join(videos_dir, uploaded_file.filename)
-
-    uploaded_file.save(video_path)
-
-    save_video(video_path)
-    save_frames(video_path)
-
-
-def get_video_paths():
-    video_paths = DB_MANAGE.query(Video.video_path).all()
-    return jsonify(path_list=[path for (path,) in video_paths])
-
